Fix for header badges and logo
In this commit, some of the links for the header badges were corrected, while other links for some others were removed (The ones that are not relevant anymore). Also, a small link pointing correction for the branding logo of the project was made. And for last, the Technical Test Devices list was updated...